# Conflicts:
#	Galaxylist/Program.cs
#	Galaxylist/client/src/lib/GalaxyView.svelte
#	Galaxylist/client/src/lib/InputFields.svelte

+++ Galaxylist/Lib/Models/Fov.cs
@@ -8,12 +8,12 @@ public class Fov
 	/// <summary>
 	/// The width of the field of view in degrees
 	/// </summary>
-	public required int Width { get; set; }
+	public required double Width { get; set; }
 
 	/// <summary>
 	/// The height of the field of view in degrees
 	/// </summary>
-	public required int Height { get; set; }
+	public required double Height { get; set; }
 
 	/// <summary>
 	/// The position angle of the field of view in degrees where 0 means the north side of the camera is perpendicular to the zenith
